fix(ui): satisfy React Doctor checks
